requirements.txt
jedi>=0.16
llama-index-core
llama-index-readers-file
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-llms-replicate
langchain
openai
langchain-openai
langchain-chroma
langchain-experimental
unstructured[all-docs]
pillow
pydantic
lxml
matplotlib
chromadb
tiktoken
llama-cpp-python
open_clip_torch
sentence-transformers
transformers
huggingface_hub
numpy
streamlit